requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
attrs==24.2.0
colorama==0.4.6
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
packaging==24.2
pluggy==1.5.0
propcache==0.2.0
pytest==8.3.3
yarl==1.17.1
